Remove id int casting

+++ src/LaravelResource/Http/Controllers/NestedResourceController.php
@@ -46,8 +46,8 @@ abstract class NestedResourceController extends BaseController
         $args = func_get_args();
         $count = count($args);
 
-        $parentId = (int)$args[$count - 2];
-        $id = (int)$args[$count - 1];
+        $parentId = $args[$count - 2];
+        $id = $args[$count - 1];
 
         $object = $this->repository->getForParent($id, $parentId);
 
@@ -75,7 +75,7 @@ abstract class NestedResourceController extends BaseController
         $args = func_get_args();
         $count = count($args);
 
-        $parentId = (int)$args[$count - 1];
+        $parentId = $args[$count - 1];
 
         $query = $this->repository->queryForParent($parentId);
 
@@ -124,8 +124,8 @@ abstract class NestedResourceController extends BaseController
         $args = func_get_args();
         $count = count($args);
 
-        $parentId = (int)$args[$count - 2];
-        $id = (int)$args[$count - 1];
+        $parentId = $args[$count - 2];
+        $id = $args[$count - 1];
 
         $object = $this->repository->getForParent($id, $parentId);
 
@@ -157,7 +157,7 @@ abstract class NestedResourceController extends BaseController
         array_shift($args);
         $count = count($args);
 
-        $parentId = (int)$args[$count - 1];
+        $parentId = $args[$count - 1];
 
         $validator = $this->validator ? $this->validator->forStore($request->all(), $args) : null;
 
@@ -206,8 +206,8 @@ abstract class NestedResourceController extends BaseController
         array_shift($args);
         $count = count($args);
 
-        $parentId = (int)$args[$count - 2];
-        $id = (int)array_pop($args);
+        $parentId = $args[$count - 2];
+        $id = array_pop($args);
 
         $validator = $this->validator ? $this->validator->forUpdate(['id' => $id] + $request->all(), $args) : null;
